requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
scikit-learn>=1.2.2
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.14.1
pandas>=2.0.2
numpy>=1.24.3
Pillow>=9.5.0
streamlit>=1.23.1
opencv-python>=4.7.0
wandb>=0.15.4
omegaconf>=2.3.0
tqdm>=4.65.0
bitsandbytes>=0.39.0
accelerate>=0.20.3
# Deployment dependencies
onnx>=1.14.0
onnxruntime-gpu>=1.15.0
onnx-simplifier>=0.4.31
# Optional: TensorRT dependencies (install separately via pip)
# nvidia-tensorrt>=8.6.1
# pycuda>=2022.2
# MLOps dependencies
mlflow>=2.3.0
dvc>=3.0.0
dvc[gdrive]>=3.0.0
dvc[s3]>=3.0.0
boto3>=1.28.0
pyyaml>=6.0
# CI/CD dependencies
pytest>=7.3.1
pytest-cov>=4.1.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0 